requirements.txt
colorama==0.4.6
Faker==40.1.2
iniconfig==2.3.0
mysql-connector-python==9.5.0
numpy==2.4.1
packaging==25.0
pandas==2.3.3
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.3
